Initialize the actual AspectJ worker only when needed.
Previously, we hold the instance in a `T.worker` which will always be initialized by mill, even if the compile target is up-to-date.
Now, we only run the initialization, when the outer task (e.g. compile) actually needs the worker.
Please note, once initialized, the worker will be still cached, as it is keept by the AspectjWorkerModule, which is a shared external module.