Support compiler chaining (weave compiled classes)
Added `aspectjCompileMode` to weave-compile after another compiler.

`CompileMode.FullSources` is the default and only used AspectJ compiler
for all compilations.

`CompileMode.OnlyAjSources` is the new setup to use another compiler to
compile the Java sources, and use AspectJ compiler only to weave-compile
those compiled classes. Only aj-files are fed as source files, all other
input (the already compiled classes) are fed via the `-inpath` option.

With this setup, we can even compile-time weave Scala (and probably also
Kotlin) classes.

For some reasons, scala+ajc does not work for Mill <=0.9.

I think it's not worth to debug it. It's probably related to some
refactoring in 0.10 regarding the "override" detection and the out
directory layout of overridden targets. But mill-aspectj still work
for older Mill versions, so we still support these platforms.